Windows Action CI (#713)
## Public-Facing Changes
Add windows CI for the C++ library. This ensures that our C++ library builds and tests pass on windows.

Closes #716

+++ cpp/test/unit_tests.cpp
@@ -79,7 +79,7 @@ TEST_CASE("internal::crc32", "[writer]") {
   };
 
   std::array<uint8_t, 32> data;
-  std::iota(data.begin(), data.end(), 1);
+  std::iota(data.begin(), data.end(), (uint8_t)(1));
 
   REQUIRE(crc32(data.data(), 0) == 0);
   REQUIRE(crc32(data.data(), 1) == 2768625435);
